Update README.md
Change single quote(') to double quote(").  It might work with single quote(didn't test) but the existing config uses double quotes.

Remove "s" in private-keys.  

Change Last pair to First pair. --- depending on how you look at it of course(as a stack or as a list), but the signing pair in the set of pairs returned from `dump_private_keys` is the first pair, not the last.  At least it was for me in following these directions.